Remove timeout on Manual On-Hold feature
- From now on, the On-Hold option will appear within visitor Info panel, provided the agent has sent the last message
