+++ packages/rocketchat-i18n/i18n/en.i18n.json
@@ -2410,6 +2410,7 @@
   "Livechat_agents": "Omnichannel agents",
   "Livechat_Agents": "Agents",
   "Livechat_allow_manual_on_hold": "Allow agents to manually place chat On Hold",
+  "Livechat_allow_manual_on_hold_description": "If enabled, the agent will get a new option to place a chat On Hold, provided the agent has sent the last message",
   "Livechat_AllowedDomainsList": "Livechat Allowed Domains",
   "Livechat_Appearance": "Livechat Appearance",
   "Livechat_auto_close_on_hold_chats_custom_message": "Custom Message to be sent when a room in On-Hold queue gets automatically closed by the system",
@@ -2437,7 +2438,6 @@
   "Livechat_last_chatted_agent_routing_Description": "The Last-Chatted Agent setting allocates chats to the agent who previously interacted with the same visitor if the agent is available when the chat starts.",
   "Livechat_managers": "Omnichannel managers",
   "Livechat_Managers": "Managers",
-  "Livechat_manual_on_hold_timeout": "How long an agent has to wait before placing the chat on-hold (in seconds), provided the agent has sent the last message",
   "Livechat_message_character_limit": "Livechat message character limit",
   "Livechat_monitors": "Livechat monitors",
   "Livechat_Monitors": "Monitors",
