+++ server/modules/watchers/publishFields.ts
@@ -65,7 +65,6 @@ export const roomFields = {
 	prid: 1,
 	avatarETag: 1,
 	usersCount: 1,
-	canPlaceChatOnHold: 1,
 
 	// @TODO create an API to register this fields based on room type
 	livechatData: 1,
